refactor: :truck: create a new `paths.py` helper module
